Crate pallet_pools

Re-exports

pub use pallet::*;

pub use weights::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_pools

Structs

EpochDetails

The result of epoch execution of a given tranch within a pool

EpochExecutionInfo

The information for a currently executing epoch

EpochExecutionTranche

EpochExecutionTranches

EpochState

HealthySolution

OutstandingCollections

The outstanding collections for an account

PoolChanges

PoolDepositInfo

Information about the deposit that has been taken to create a pool

PoolDetails

PoolParameters

ReserveDetails

ScheduledUpdateDetails

Tranche

TrancheInput

TrancheMetadata

TrancheSolution

TrancheUpdate

Tranches

UnhealthySolution

UserOrder

Per-tranche and per-user order details.

Enums

EpochSolution

The solutions struct for epoch solution

PoolState

PoolStatus

TrancheLoc

TrancheType

UnhealthyState

Functions

calculate_solution_parameters

Type Definitions

Seniority

Type that indicates the seniority of a tranche

TrancheIndex

The index type for tranches

TrancheSalt

The salt type for tranches