Trait pallet_pools::pallet::Config

pub trait Config: Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen + FixedPointOperand + From<u64> + From<u128> + TypeInfo + TryInto<u64>;
    type TrancheWeight: Parameter + Copy + Convert<Self::TrancheWeight, Self::Balance> + From<u128>;
    type BalanceRatio: Member + Parameter + Default + Copy + TypeInfo + FixedPointNumber<Inner = Self::Balance>;
    type InterestRate: Member + Parameter + Default + Copy + TypeInfo + FixedPointNumber<Inner = Self::Balance>;
    type PalletId: Get<PalletId>;
    type PalletIndex: Get<u8>;
    type PoolId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen + Debug;
    type TrancheId: Member + Parameter + Default + Copy + MaxEncodedLen + TypeInfo + From<[u8; 16]>;
    type EpochId: Member + Parameter + Default + Copy + AtLeast32BitUnsigned + HasCompact + MaxEncodedLen + TypeInfo + Into<u32>;
    type CurrencyId: Parameter + Copy;
    type PoolCurrency: Contains<Self::CurrencyId>;
    type UpdateGuard: PoolUpdateGuard<PoolDetails = PoolDetails<<Self as Config>::CurrencyId, <Self as Config>::EpochId, <Self as Config>::Balance, <Self as Config>::InterestRate, <Self as Config>::MaxSizeMetadata, <Self as Config>::TrancheWeight, <Self as Config>::TrancheId, <Self as Config>::PoolId>, ScheduledUpdateDetails = ScheduledUpdateDetails<<Self as Config>::InterestRate, <Self as Config>::MaxTokenNameLength, <Self as Config>::MaxTokenSymbolLength, <Self as Config>::MaxTranches>, Moment = Moment>;
    type AssetRegistry: OrmlMutate<AssetId = Self::CurrencyId, Balance = Self::Balance, CustomMetadata = CustomMetadata>;
    type ParachainId: Get<ParachainId>;
    type Currency: ReservableCurrency<Self::AccountId, Balance = Self::Balance>;
    type Tokens: Mutate<Self::AccountId> + Inspect<Self::AccountId, AssetId = Self::CurrencyId, Balance = Self::Balance> + Transfer<Self::AccountId>;
    type Permission: Permissions<Self::AccountId, Scope = PermissionScope<Self::PoolId, Self::CurrencyId>, Role = Role<Self::TrancheId, Moment>, Error = DispatchError>;
    type NAV: PoolNAV<Self::PoolId, Self::Balance>;
    type TrancheToken: TrancheToken<Self::PoolId, Self::TrancheId, Self::CurrencyId>;
    type Time: UnixTime;
    type ChallengeTime: Get<<Self as Config>::BlockNumber>;
    type DefaultMinEpochTime: Get<u64>;
    type DefaultMaxNAVAge: Get<u64>;
    type MinEpochTimeLowerBound: Get<u64>;
    type MinEpochTimeUpperBound: Get<u64>;
    type MaxNAVAgeUpperBound: Get<u64>;
    type MinUpdateDelay: Get<u64>;
    type MaxSizeMetadata: Get<u32> + Copy + Member + TypeInfo;
    type MaxTokenNameLength: Get<u32> + Copy + Member + TypeInfo;
    type MaxTokenSymbolLength: Get<u32> + Copy + Member + TypeInfo;
    type MaxTranches: Get<u32> + Member + TypeInfo;
    type PoolDeposit: Get<Self::Balance>;
    type PoolCreateOrigin: EnsureOrigin<Self::Origin>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen + FixedPointOperand + From<u64> + From<u128> + TypeInfo + TryInto<u64>

type TrancheWeight: Parameter + Copy + Convert<Self::TrancheWeight, Self::Balance> + From<u128>

type BalanceRatio: Member + Parameter + Default + Copy + TypeInfo + FixedPointNumber<Inner = Self::Balance>

A fixed-point number which represents the value of one currency type in terms of another.

type InterestRate: Member + Parameter + Default + Copy + TypeInfo + FixedPointNumber<Inner = Self::Balance>

A fixed-point number which represents an interest rate.

type PalletId: Get<PalletId>

type PalletIndex: Get<u8>

The immutable index of this pallet when instantiated within the context of a runtime where it is used.

type PoolId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen + Debug

type TrancheId: Member + Parameter + Default + Copy + MaxEncodedLen + TypeInfo + From<[u8; 16]>

type EpochId: Member + Parameter + Default + Copy + AtLeast32BitUnsigned + HasCompact + MaxEncodedLen + TypeInfo + Into<u32>

type CurrencyId: Parameter + Copy

type PoolCurrency: Contains<Self::CurrencyId>

type UpdateGuard: PoolUpdateGuard<PoolDetails = PoolDetails<<Self as Config>::CurrencyId, <Self as Config>::EpochId, <Self as Config>::Balance, <Self as Config>::InterestRate, <Self as Config>::MaxSizeMetadata, <Self as Config>::TrancheWeight, <Self as Config>::TrancheId, <Self as Config>::PoolId>, ScheduledUpdateDetails = ScheduledUpdateDetails<<Self as Config>::InterestRate, <Self as Config>::MaxTokenNameLength, <Self as Config>::MaxTokenSymbolLength, <Self as Config>::MaxTranches>, Moment = Moment>

type AssetRegistry: OrmlMutate<AssetId = Self::CurrencyId, Balance = Self::Balance, CustomMetadata = CustomMetadata>

type ParachainId: Get<ParachainId>

type Currency: ReservableCurrency<Self::AccountId, Balance = Self::Balance>

type Tokens: Mutate<Self::AccountId> + Inspect<Self::AccountId, AssetId = Self::CurrencyId, Balance = Self::Balance> + Transfer<Self::AccountId>

type Permission: Permissions<Self::AccountId, Scope = PermissionScope<Self::PoolId, Self::CurrencyId>, Role = Role<Self::TrancheId, Moment>, Error = DispatchError>

type NAV: PoolNAV<Self::PoolId, Self::Balance>

type TrancheToken: TrancheToken<Self::PoolId, Self::TrancheId, Self::CurrencyId>

A conversion from a tranche ID to a CurrencyId

type Time: UnixTime

type ChallengeTime: Get<<Self as Config>::BlockNumber>

Challenge time

type DefaultMinEpochTime: Get<u64>

Pool parameter defaults

type DefaultMaxNAVAge: Get<u64>

type MinEpochTimeLowerBound: Get<u64>

Pool parameter bounds

type MinEpochTimeUpperBound: Get<u64>

type MaxNAVAgeUpperBound: Get<u64>

type MinUpdateDelay: Get<u64>

Pool update settings

type MaxSizeMetadata: Get<u32> + Copy + Member + TypeInfo

Max size of Metadata

type MaxTokenNameLength: Get<u32> + Copy + Member + TypeInfo

Max length for a tranche token name

type MaxTokenSymbolLength: Get<u32> + Copy + Member + TypeInfo

Max length for a tranche token symbol

type MaxTranches: Get<u32> + Member + TypeInfo

Max number of Tranches

type PoolDeposit: Get<Self::Balance>

The amount that must be reserved to create a pool

type PoolCreateOrigin: EnsureOrigin<Self::Origin>

The origin permitted to create pools

type WeightInfo: WeightInfo

Weight Information